[FIX] Insert intermediate variables to store values from ENVs
If a existing variable is used inside a condition and with a ":=" statetment, this will be considered a variable inside this context and its correspondent value won't be stored in the original variable. In that case, we'll use an intermediate variable to store that value so we can recover this from the original variable.